refactor: Implement usePostArticle hook
