+++ client/src/Hooks/index.js
@@ -3,6 +3,7 @@ import useLogout from "./useLogout"
 import useGetSidebarContent from "./useGetSidebarContent"
 import useGetArticle from "./useGetArticle"
 import useConfirmEmail from "./useConfirmEmail"
+import usePostArticle from "./usePostArticle"
 
 export{
     useAuth,
@@ -10,4 +11,5 @@ export{
     useGetSidebarContent,
     useGetArticle,
     useConfirmEmail,
+    usePostArticle,
 }
